requirements.txt
biopython>=1.81
matplotlib>=3.5.0
numpy>=1.21.0 